TEST: Separated classes in different files.
