+++ dbod/tests/rundeck.py
@@ -1,0 +1,35 @@
+
+# Copyright (C) 2015, CERN
+# This software is distributed under the terms of the GNU General Public
+# Licence version 3 (GPL Version 3), copied verbatim in the file "COPYING".
+# In applying this license, CERN does not waive the privileges and immunities
+# granted to it by virtue of its status as Intergovernmental Organization
+# or submit itself to any jurisdiction.
+
+import unittest
+from types import *
+import json
+import logging
+import sys
+import requests
+logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
+
+class Rundeck(unittest.TestCase):
+    @classmethod
+    def setUpClass(self):
+        pass
+
+    @classmethod
+    def tearDownClass(self):
+        pass
+        
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+        
+    def test_connection(self):
+        response = requests.get("https://localhost:5432/", verify=False)
+        self.assertEquals(response.status_code, 200)
+
